Retrieve sound files from CDSTAR
This commit adapts the JSON data delivered to the frontend such that
CDSTAR URLs are included instead of relative URLs on the app's site.

It requires the CDSTAR catalog data to be present in a table "soundfiles",
which needs to be created via the corresponding fabric task
"load_soundfile_catalog".